Fixed memory leak with circle drawing
SDL_RenderDrawPoint seemed to have a memory issue which was more
apparent with circle drawing as it called it multiple times. To solve
this SDL_RenderDrawPoints was used to reduce function calls.

Apparently using a stack array calling the render function multiple
times is faster than appending a list of points to a vector and
rendering once after the loop. Further tests needed.